fungibility: retimer app to update sku id and software id
For fungibility software Id and SKU property needs to be updated. Since
it's not a PLDM device it should be done as part of seperate
application.

This patch adds service to update the D-Bus properties for retimer. SKU
property is formed by concatinating device and vendor id for retimer and
software id will be updated with component identifier value for retimer.

Unit Tests:
  1. Verify software id is set with component identifier value for
     retimer
  2. Verify SKU property is set by concatinating device and vendor id
     with fixed size of 4 bytes
  3. Check nvidia-retimer-app is run once during bootup and updates when
     objects are populated by gpu manager.
  4. restart gpu manager and check retimer properties are populated once gpu
     manager populates the object
  5. Restart nvidia retimer app and retimer properties are updated

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-5032

Signed-off-by: Chinmay Shripad Hegde <chinmays@nvidia.com>